[refactor/#224] my-reservatios-responses.ts 파일명 오타 수정 및 camelCase로 변경, import 경로 수정
