+++ src/components/MyReservations/ReservationCard.tsx
@@ -4,7 +4,7 @@ import ReviewModal from "@/src/components/Modal/ReviewModal/ReviewModal";
 import RESERVATION_LABEL from "@/src/constants/reserbationStatus";
 import { usePatchMyReservation } from "@/src/queries/useMyReservations";
 import useModalStore from "@/src/stores/ModalStore";
-import { Reservation } from "@/src/types/my-reservatios-responses";
+import { Reservation } from "@/src/types/myReservationsResponses";
 import { formatDate3 } from "@/src/utils/calendarFormatDate";
 import clsx from "clsx";
 import Image from "next/image";
